requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
websockets==12.0
langchain==0.1.0
langchain-google-genai==0.0.6
jinja2==3.1.2 